requirements.txt
Django==2.2.3
django-autoslug==1.9.4
django-ckeditor==5.7.1
django-js-asset==1.2.2
djangorestframework==3.10.2
Pillow==6.1.0
pytz==2019.1
sorl-thumbnail==12.5.0
sqlparse==0.3.0
